A proposed change to the Peeker class.
Proposing a change where a `Peeker` object can be created with a set of
signals instead of a `Peeker` foreach.

```python
p = Peeker(z=z, a=a, b=b, sel=sel)
p.to_wavedrom()
```

This creates a `Peeker` object with the signals to be "peeked" at, in the
keyword parameters the names can be specified.  This should work the same
as the existing but you don't need to create a `Peeker` for each signal,
just a `Peeker` for a set of signals.  With this approach you can also
create various `Peeker`s at different levels.

Things that should be added:

  `Peeker.add_signal(sig)``: method that allows a signal to be added after
    creation, could even overload the `__add__`; `p += mysig`

  `Peeker.rm_signal(name)`

  Add checkers to `to_wavejson` to verify no parameters where passed
  via `names` (\*args).