Added a merge_all function.
With the API changes one feature was missing from the previous and that
was the ability to have a single global peeker that included the signals
of all the peekers added through the design.  This is handled by keeping
track of all the keepers in a global dict.

Things that need to be added:

1. add __del__ to remove a peeker from the global list when it is deleted.
2. clear_all function to clear all the keepers from the global list.