Update package version and refactor code for compatibility with Python 3
This commit removes the use of `future` and its `standard_library` module from myhdlpeek package as Python 2 is no longer in use. Also, the method of determining the number of bits for both integer and intbv types in the 'Peeker' class has been modified to properly reflect the correct bit length in Python 3. Additionally, the package version has been updated.